APIClientID converted to int64 as well, for consistency.

+++ applicationcharge.go
@@ -26,7 +26,7 @@ type ApplicationChargeServiceOp struct {
 type ApplicationCharge struct {
 	ID                 int64            `json:"id"`
 	Name               string           `json:"name"`
-	APIClientID        int              `json:"api_client_id"`
+	APIClientID        int64              `json:"api_client_id"`
 	Price              *decimal.Decimal `json:"price"`
 	Status             string           `json:"status"`
 	ReturnURL          string           `json:"return_url"`
